requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mysql-connector-python==9.0.0
passlib[argon2]==1.7.4
pydantic==2.9.2
python-dotenv==1.0.1